% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandExpectedVal.R
\name{LandExpectVal}
\alias{LandExpectVal}
\title{Land Expected Value (Multiple Rotations)}
\usage{
LandExpectVal(FutureValue, Age, Discount)
}
\arguments{
\item{FutureValue}{Total Future Value (Use SinglePayment and SeriesPayment Functions)}

\item{Age}{The age where LEV is of interest}

\item{Discount}{The discount rate (ex: .05)}
}
\value{
Land Expected Value for a single year when the stand has multiple rotations.
}
\description{
Use this function to estimate the simple LEV
}
\examples{

Age <- 40
FutureValue <- 1960
Discount <- .06

LandExpectVal(Age, FutureValue, Discount)
}
\seealso{
Other Rotation Age: 
\code{\link{SingleRotationVal}()}
}
\concept{Rotation Age}
