% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeriesPayment.R
\name{SeriesPayment}
\alias{SeriesPayment}
\title{The present or future value of a series of payments.}
\usage{
SeriesPayment(
  Flow,
  Rate = 0.06,
  Frequency = "Annual",
  PeriodLength = NA,
  Termination = NA,
  Present = TRUE
)
}
\arguments{
\item{Flow}{Value of asset or revenue/expense.}

\item{Rate}{either the discount rate (if calculating present value) or interest rate (
for value using compounding interest)}

\item{Frequency}{"Annual" or "Periodic" payments. Defaults to "Annual"}

\item{PeriodLength}{NA or Numeric. Numeric years between periods for "Periodic" payments. NA for annual payments.}

\item{Termination}{Time horizon in years. NA for 'forever' or non-terminating payments.}

\item{Present}{TRUE or FALSE. TRUE will calculate the present value of payment made at year Time
and FALSE the future value of a payment with compounding interest in Time years.}
}
\description{
The following six possible series payments can be calculated:

\itemize{
\item Series Annual Forever Present Value:
\item Frequency = "Annual", Termination = NA, Present = TRUE }

\itemize{
\item Series Annual Terminating Present Value:
\item Frequency = "Annual", Termination = Numeric(years), Present = TRUE }

\itemize{
\item Series Annual Terminating Future Value:
\item Frequency = "Annual", Termination = numeric(years), Present = FALSE }

\itemize{
\item Series Periodic Forever Present Value:
\item Frequency = "Periodic", PeriodLength = numeric(years.between.periods),
Termination = NA, Present = TRUE }

\itemize{
\item Series Periodic Terminating Present Value:
\item Frequency = "Periodic", PeriodLength = numeric(years.between.periods),
Termination = numeric(years), Present = TRUE }

\itemize{
\item Series Periodic Terminating Future Value:
\item  Frequency = "Periodic", PeriodLength = numeric(years.between.periods),
Termination = numeric(years), Present = FALSE }
}
\examples{

SeriesPayment(100, .05, "Annual", NA, NA, TRUE)
SeriesPayment(100, .05, "Annual", NA, Termination = 30, TRUE)
SeriesPayment(100, .05, "Annual", NA, Termination = 30, FALSE)
SeriesPayment(100, .05, "Periodic", PeriodLength = 5, Termination = NA, TRUE)
SeriesPayment(100, .05, "Periodic", PeriodLength = 5, Termination = 30, TRUE)
SeriesPayment(100, .05, "Periodic", PeriodLength = 5, Termination = 30, FALSE)

}
\seealso{
Other Payment Functions: 
\code{\link{SinglePayment}()}
}
\concept{Payment Functions}
