% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleNPV.R
\name{SimpleNPV}
\alias{SimpleNPV}
\title{Net Present Value}
\usage{
SimpleNPV(Flow, Class, Occurance, Discount, NegativeValues = TRUE)
}
\arguments{
\item{Flow}{Value of asset or revenue/expense}

\item{Class}{asset or flow class use either "Expense"/"Revenue" or "Positive"/"Negative"}

\item{Occurance}{Years from today the revenue occurs. Today is year 0.}

\item{Discount}{Discount rate. Enter as .05 or .07, etc. Do not enter vector only single number.}

\item{NegativeValues}{if set to FALSE will change flow values to negative numbers for
expenses and negative assets that have not been
entered as negative numbers in the Flow vector. Expenses or negative values must
be given an Expense or Negative Class value to work. Defaults to TRUE.}
}
\value{
Net Present Value
}
\description{
This function takes in any number of positive or negative values or negative
flows to calculate NPV as a specific discount rate.
}
\details{
Run function at multiple discount rates for sensitivity analysis.
}
\examples{

Flow <- c(-400, -100, 0, 200, 6600)
Class <- c("expense", "Expense", "Revenue", "Revenue", "Revenue")
Occurance <- c(0,5,8,15,30)
Discount <- .06
NPV(Flow, Class, Occurance, Discount)

\dontrun{
Example if flows are not given negative values when they are expenses
Flow <- c(400, 100, 0, 200, 6600)
Class <- c("negative", "negative", "positive", "positive", "positive")
Occurance <- c(0,5,8,15,30)
Discount <- .06
NegativeValues == FALSE
}
}
\seealso{
Other Value Functions: 
\code{\link{NPV}()},
\code{\link{PresentValue}()}
}
\concept{Value Functions}
