% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinglePayment.R
\name{SinglePayment}
\alias{SinglePayment}
\title{Single Payment Future of Present Value}
\usage{
SinglePayment(Flow, Rate, Time, Present = TRUE)
}
\arguments{
\item{Flow}{Value of asset or revenue/expense.}

\item{Rate}{either the discount rate (if calculating present value) or interest rate (
for value using compounding interest)}

\item{Time}{Time horizon in years.}

\item{Present}{TRUE or FALSE. TRUE will calculate the present value of payment made at year Time
and FALSE the future value of a payment with compounding interest in Time years.}
}
\value{
Vector of length Flow
}
\description{
This function calculates either the present value of a future cash flow or the
compounded future value of a cash flow after n years.
}
\details{
Must be fed a matrix containing Rate, Time, and Present values for all payments.
}
\examples{

Flow <- c(rep(c(500,600,700,800,900),2))
Rate <- rep(.06, 10)
Time <- rep(30, 10)
Present <- c(rep(TRUE,5), rep(FALSE,5))
SinglePayment(Flow, Rate, Time, Present)
SinglePayment(6000, .05, 40, TRUE)
SinglePayment(6000, .05, 40, FALSE)
SinglePayment(6000, .06, 40, FALSE)

}
\seealso{
Other Payment Functions: 
\code{\link{SeriesPayment}()}
}
\concept{Payment Functions}
