% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleRotationVal.R
\name{SingleRotationVal}
\alias{SingleRotationVal}
\title{Single Rotation Value (Multiple Rotations)}
\usage{
SingleRotationVal(Age, Stumpage, Volume, Cost, Discount)
}
\arguments{
\item{Age}{The age where NPV is of interest}

\item{Stumpage}{the stumpage value in a given year}

\item{Volume}{The yield in a given year}

\item{Cost}{The total cost expanded to a given year cost (ie: regeneration costs or property taxes)}

\item{Discount}{The discount rate (ex: .05)}
}
\value{
The NPV at a single age for a stand when deciding optimal rotation.
}
\description{
Use this function to estimate the simple single rotation value.
}
\examples{
Age <- 40
Stumpage <- 300
Volume <- 20
Cost <- 100
Discount <- .06

SingleRotationVal(75, 150, 30, 100, .06)


}
\seealso{
Other Rotation Age: 
\code{\link{LandExpectVal}()}
}
\concept{Rotation Age}
