% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexNPV.R
\name{ComplexNPV}
\alias{ComplexNPV}
\title{ComplexNPV}
\usage{
ComplexNPV(
  Flow,
  Class,
  Frequency,
  FirstOccurance,
  TimeHorizon,
  Rate,
  PeriodLength = NA,
  Discount
)
}
\arguments{
\item{Flow}{cash flow/asset/expense of interest.}

\item{Discount}{Discount rate. Enter percent as .05 or .07, etc. Do not enter vector only single number.}

\item{Year}{age at which you want to know flow's value.}
}
\value{
Present Value
}
\description{
This function calculates a single asset's value
at a specific time at a single discount rate.
}
\details{
Run function at multiple discount rates for sensitivity analysis.
}
\examples{

PresentValue(100000, .06, 30)
PresentValue(100000, .06, 20)
PresentValue(100000, .07, 30)
PresentValue(100000, .07, 20)

}
\seealso{
Other Value Functions: 
\code{\link{PresentValue}()},
\code{\link{SimpleNPV}()}
}
\concept{Value Functions}
